Initial implementation of python-ecies library
Python implementation of ECIES (Elliptic Curve Integrated Encryption Scheme).

This library implements ECIES in a configurable matter where the key derivation
function, the encryption and the MAC functions as well as the output format can
be customized.

The library itself provides the following functionality:

Key exchange:
 - ECDH (normal)

Key derivation:
 - HKDF (RFC 5869)
 - ConcatKDF (NIST SP 800-56Ar3)
 - X9.63-KDF

Encryption & MAC:
 - AES-GCM (AEAD)
 - AES-128/256/192 with any HMAC
 - XOR with any HMAC

Output format:
 - Short binary
 - JSON with PEM ephemeral key

The code is in the following state:

 - Built on top of python cryptography library with optional HSM support
   via YubiHSM library

 - Linting, testing and documentation generation via tox

 - ruff linting, formating and mypy clean

 - Easily extendable, and well documented formats to allow porting and
   interoperability with other languages.

 - Basic getting-started documentation